backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-generativeai==0.3.2
pydantic==2.6.1 